campus_checkin_platform/requirements.txt

Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
django-environ==0.11.2


djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0


mysqlclient==2.2.0


redis==5.0.1
django-redis==5.3.0
celery==5.3.4
django-celery-beat==2.5.0


channels==4.0.0
channels-redis==4.1.0


Pillow==10.0.0


python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2


gunicorn==21.2.0
whitenoise==6.6.0


pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0


drf-yasg==1.21.7

pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0